Stub out marshalling of the private RSpec Time class copy
Recent versions of RSpec clone the Time class, to provide access to `now`
without having to worry about stubs left over from tests.  This led to a
failure in TorqueSpec when an attempt was made to marshall those objects.

This provides stubs that simply ship nothing -- no state -- for those
instances.  While this probably isn't optimal, it is sufficient to make remote
tests work as expected with RSpec 2.13

Signed-off-by: Daniel Pittman <daniel@rimspace.net>