Define `pending_fixed?` on exceptions shipped across the wire
TorqueSpec translates all exceptions thrown inside the container into
`Exception` instances, to avoid dependencies on classes that may not be
present on the client side.

Unfortunately, this has the side-effect of breaking detection of RSpec pending
tests that suddenly began to work -- a feature of newer versions.

This augments the marshalling to ship that extra field across the wire, and
attaches it appropriate exceptions; this gets RSpec 2.13 working again.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>